POL-68: refleact review
