+++ gatsby-node.js
@@ -23,6 +23,26 @@ exports.createPages = async ({actions, graphql}) => {
         })
     })
 
+    //委員会詳細ページ
+    const committeesResult = await graphql(`
+    {
+        politylink {
+            Committee {
+                id
+            }
+        }
+    }
+    `)
+    committeesResult.data.politylink.Committee.forEach(({id}) => {
+        createPage({
+            path: "/committee/" + id.split(':').pop(),
+            component: path.resolve(`./src/templates/committee.js`),
+            context: {
+                committeeId: id,
+            }
+        })
+    })
+
     // 会議録詳細ページ
     const minutesResult = await graphql(`
     {
