+++ src/pages/committees.js
@@ -51,7 +51,7 @@ export default class App extends React.Component {
                     {filteredCommittees.map((committees) => {
                         return <CommitteeCard
                             title={committees.name}
-                            to={"/committees/" + committees.id.split(':').pop()}
+                            to={"/committee/" + committees.id.split(':').pop()}
                         />;
                     })}
                 </FlexContainer>
