+++ src/templates/minutes.js
@@ -2,35 +2,28 @@ import React from "react"
 import {graphql} from "gatsby"
 import styles from "./minutes.module.css"
 import {Container, FlexContainer} from "../components/container"
-import Layout from "../components/layout";
-import SEO from "../components/seo";
-import LinkCard from "../components/linkCard";
-import BillCard from "../components/billCard";
+import Layout from "../components/layout"
+import SEO from "../components/seo"
+import LinkCard from "../components/linkCard"
+import BillCard from "../components/billCard"
+import {formatDate, formatSentence} from "../utils/format"
 
-
-export const formatStartDate = (date) => {
-    return String(date.year) + "/" + String(date.month).padStart(2, '0') + "/" + String(date.day).padStart(2, '0')
-}
-
-export const formatTopic = (topic) => {
-    return "- " + topic
-}
 
 export default function Minutes({data}) {
     const minutes = data.politylink.Minutes[0]
-    const description = formatStartDate(minutes.startDateTime) + "開催の" + minutes.name + "に関する情報をまとめています。"
+    const description = formatDate(minutes.startDateTime) + "開催の" + minutes.name + "に関する情報をまとめています。"
 
     return (
         <Layout>
             <SEO title={minutes.name} description={description}/>
             <Container>
                 <h2 className={styles.name}>{minutes.name}</h2>
-                <h3 className={styles.number}>{formatStartDate(minutes.startDateTime)}</h3>
+                <h3 className={styles.number}>{formatDate(minutes.startDateTime)}</h3>
                 <div className={styles.summary}>
                     <p>{minutes.summary}</p>
                     <Container>
                         {minutes.topics.filter(topic => topic != null).map((topic) => {
-                            return <p className={styles.topic}> {formatTopic(topic)} </p>
+                            return <p className={styles.topic}> {formatSentence(topic)} </p>
                         })}
                     </Container>
                 </div>
