+++ src/utils/format.js
@@ -1,0 +1,7 @@
+export const formatSentence = (sentence) => {
+    return "- " + sentence
+}
+
+export const formatDate = (date) => {
+    return String(date.year) + "/" + String(date.month).padStart(2, '0') + "/" + String(date.day).padStart(2, '0')
+}
